attempt 3
So I added a widget that provides the instagram feed. It does feature the widget's creator. I don't know why the icons won't show up in the centre.  The contact overall does look a little weird; the ratios are not there yet (between instagram feed & the form) and I still want to play with the fonts. Overall this is just the basic.